Refactor scheduling code
