+++ src/game/themes/index.ts
@@ -1,4 +1,5 @@
 import { levelModifiers } from "@/game/level-types/types";
+import { filterInRange, RangedItem } from "@/support/schedule";
 
 import { BlockColor } from "../types";
 
@@ -54,13 +55,6 @@ export const getColorMapping = (
   return mapping[theme];
 };
 
-type ScheduleDate = { month: number; day: number };
-
-export type RangedItem = {
-  begin: ScheduleDate;
-  end: ScheduleDate;
-};
-
 export type ThemeSchedule = RangedItem & {
   name: string;
   theme: BlockTheme;
@@ -110,36 +104,6 @@ export const themeSchedule: ThemeSchedule[] = [
   }
 ];
 
-const inRange = (date: Date, begin: ScheduleDate, end: ScheduleDate) => {
-  const day = date.getDate();
-  const month = date.getMonth() + 1;
-
-  const afterStart =
-    begin.month < month || (begin.month === month && begin.day <= day);
-
-  const beforeEnd =
-    end.month > month || (end.month === month && end.day >= day);
-
-  return afterStart && beforeEnd;
-};
-
-const filterInRange = <TItem extends RangedItem>(
-  date: Date,
-  items: TItem[]
-): TItem[] =>
-  items.filter((schedule) => {
-    if (schedule.begin.month > schedule.end.month) {
-      // Schedule spans over the year, split it into two ranges
-      return (
-        inRange(date, schedule.begin, EOY) || inRange(date, BOY, schedule.end)
-      );
-    }
-    return inRange(date, schedule.begin, schedule.end);
-  });
-
-const EOY: ScheduleDate = { month: 12, day: 31 };
-const BOY: ScheduleDate = { month: 1, day: 1 };
-
 export const getActiveTheme = (date: Date) => {
   const activeSchedule = filterInRange(date, themeSchedule)[0];
   return activeSchedule?.theme ?? "default";
@@ -149,14 +113,3 @@ export const getActiveModifiers = (date: Date) =>
   filterInRange(date, themeSchedule).flatMap((theme) =>
     theme.levelModifiers ? filterInRange(date, theme.levelModifiers) : []
   );
-
-/**
- * Get the current date.
- *
- * The app uses this function to determine the active theme.
- * Dates can be overridden for testing purposes.
- *
- * @returns the current date
- */
-export const getToday = (): Date =>
-  process.env.NODE_ENV === "production" ? new Date() : new Date();
