+++ src/modules/Help/index.tsx
@@ -6,7 +6,8 @@ import { DialogTitle } from "@/ui/Dialog/DialogTitle";
 import { ReactComponent as Christmas } from "@/../docs/christmas.md";
 import { ReactComponent as Halloween } from "@/../docs/halloween.md";
 import { ReactComponent as HelpText } from "@/../docs/help.md";
-import { getActiveTheme, getToday } from "@/game/themes";
+import { getActiveTheme } from "@/game/themes";
+import { getToday } from "@/support/schedule";
 
 import styles from "../Settings/TextStyling.module.css";
